Seems that florence is taken the autohide action passing a configuration file with autohide=false This break the lxdm login. The patch enable the autohide mode only if is not passed the no-gnome option. I use this option at login time and then I can jump the autohide switch mode that hide the keyboard at the startup.
This is only a rough patch I would see the keyboard at startup
if is specified HIDE_ON_START=FALSE and AUTO_HIDE=TRUE
in both cases gsettings and con file.

The desired behaviour is that the gsettings settings are neglected
if a configuration file has been specified by command line.